scanner-core/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-nmap==0.7.1
dnspython==2.4.2
urllib3==2.1.0
lxml==4.9.3
selenium==4.16.0
aiohttp==3.9.1
colorama==0.4.6
python-dotenv==1.0.0
pyyaml==6.0.1
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
packaging==23.2
reportlab==4.0.0